\name{zmat}
\alias{zmat}
\alias{SNPlocations}
\docType{data}
\title{
Three GWAS studies 
}
\description{
The data of the three studies was simulated with HAPGEN2 and a sample of it was taken (Chromosomes 1 to 4) for the examples. The data is summarized as z-scores (transformed p-values, with inverse standard normal cumulative distribution). 
}
\usage{data(zmat)}
\format{Matrix of 249024 SNPs' z-scores (in rows) in each of the 3 studies (columns).
}
\source{
See:
Su, Zhan, Jonathan Marchini, and Peter Donnelly. "HAPGEN2: simulation of multiple disease SNPs." Bioinformatics 27.16 (2011): 2304-2305.
}
\examples{
data(zmat)
}
\keyword{datasets}
\keyword{Replicability}
\keyword{GWAS}
